can train now starting from a specific ckpt

+++ examples/get_up/figure_train.py
@@ -17,22 +17,29 @@ print("Using device:", device)
 
 def main():
     parser = argparse.ArgumentParser()
-    parser.add_argument("-e", "--exp_name", type=str, default="getup")
+    parser.add_argument("-e", "--exp_name", type=str, required=True)
     parser.add_argument("-B", "--num_envs", type=int, default=4096)
     parser.add_argument("--max_iterations", type=int, default=500)
     parser.add_argument("-v", "--vis", action="store_true", default=False)
+    parser.add_argument("--ckpt", type=int, default=-1)
     args = parser.parse_args()
 
     backend = gs.constants.backend.cpu if device == "cpu" else gs.constants.backend.gpu
     gs.init(logging_level="info", backend=backend)
 
-    log_dir = f"logs/{args.exp_name}/{datetime.now().strftime('%Y%m%d_%H%M%S')}/"
-    env_cfg, obs_cfg, reward_cfg = get_cfgs()
-    train_cfg = get_train_cfg(args.exp_name, args.max_iterations)
+    assert args.exp_name is not "", "Please provide an experiment name"
 
-    if os.path.exists(log_dir):
-        shutil.rmtree(log_dir)
-    os.makedirs(log_dir, exist_ok=True)
+    if args.ckpt != -1: # resume training
+        log_dir = f"logs/{args.exp_name}/"
+        env_cfg, obs_cfg, reward_cfg, train_cfg = pickle.load(open(f"logs/{args.exp_name}/cfgs.pkl", "rb"))
+    else:
+        log_dir = f"logs/{args.exp_name}/{datetime.now().strftime('%Y%m%d_%H%M%S')}/"
+        env_cfg, obs_cfg, reward_cfg = get_cfgs()
+        train_cfg = get_train_cfg(args.exp_name, args.max_iterations)
+
+        if os.path.exists(log_dir):
+            shutil.rmtree(log_dir)
+        os.makedirs(log_dir, exist_ok=True)
 
     writer = SummaryWriter(log_dir=log_dir, flush_secs=10)
 
@@ -40,12 +47,22 @@ def main():
         num_envs=args.num_envs, env_cfg=env_cfg, obs_cfg=obs_cfg, reward_cfg=reward_cfg, device=device, writer=writer, show_viewer=args.vis,
     )
 
+    print("Creating runner")
     runner = OnPolicyRunner(env, train_cfg, log_dir, device=device)
 
-    pickle.dump(
-        [env_cfg, obs_cfg, reward_cfg, train_cfg],
-        open(f"{log_dir}/cfgs.pkl", "wb"),
-    )
+    # If training is resumed, load the runner
+    if args.ckpt != -1:
+        print("Loading model")
+        resume_path = os.path.join(log_dir, f"model_{args.ckpt}.pt")
+        runner.load(resume_path)
+
+    # If training is not resumed, save the configurations
+    if args.ckpt == -1:
+        print("Saving configurations")
+        pickle.dump(
+            [env_cfg, obs_cfg, reward_cfg, train_cfg],
+            open(f"{log_dir}/cfgs.pkl", "wb"),
+        )
 
     init_at_random_ep_len = False
